test(ui): colocate bulk-delete lifecycle coverage
